00_tutorial_snap/requirements.txt

snap-stanford


